feat(strategy): add screen space pruning threshold and densification info
add max_radii_screen field to DensificationInfo to track gaussian sizes
change max_screen_size from int to float for relative thresholding
update pruning logic to consider both world and screen space sizes
modify forward pass to compute and store screen space radii
adjust default config values for better performance

+++ tinygs/src/strategy/strategy.cpp
@@ -97,7 +97,7 @@ void StrategyParams::from_json(const json& config) {
     pruning_scale_threshold = config["pruning_scale_threshold"].get<float>();
   }
   if (config.contains("max_screen_size")) {
-    max_screen_size = config["max_screen_size"].get<int>();
+    max_screen_size = config["max_screen_size"].get<float>();
   }
   if (config.contains("duplicate_grad_threshold")) {
     duplicate_grad_threshold = config["duplicate_grad_threshold"].get<float>();
